+++ packages/datatrak-web/src/features/Tasks/TasksTable/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+
+export { TasksTable } from './TasksTable';
